Add highlight on hover functionality
Was adding event listeners to square divs before appending them to the container div. Need to append first, then add listeners. For each loop was not working. Could not use method getElementById on the squares as they were all identical and so had to add event listeners as they were created, in each iteration of the for loop.